+++ application/SaLsA/de.offis.salsa.playground/src/de/uniol/inf/is/odysseus/salsa/ui/GridScreen.java
@@ -37,7 +37,7 @@ public class GridScreen extends JFrame {
         this.setDefaultCloseOperation(HIDE_ON_CLOSE);
     }
 
-    public void onGrid(final Byte[][] grid) {
+    public void onGrid(final Double[][] grid) {
         this.map.onGrid(grid);
     }
 }
